requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
ddgs>=1.0.0

